+++ src/model/Misc.h
@@ -27,6 +27,7 @@
 #include <QString>
 #include <QSqlRecord>
 
+#include "model/FolderPropertyBase.h"
 #include "model/Ingredient.h"
 #include "model/IngredientBase.h"
 #include "model/IngredientAmount.h"
@@ -71,7 +72,8 @@ class Misc : public Ingredient,
    INGREDIENT_BASE_DECL(Misc)
    FOLDER_BASE_DECL(Misc)
    // See model/FolderPropertyBase.h for info, getters and setters for these properties
-   Q_PROPERTY(int containedInFolderId   READ containedInFolderId   WRITE setContainedInFolderId)
+   Q_PROPERTY(int     containedInFolderId     READ containedInFolderId     WRITE setContainedInFolderId  )
+   Q_PROPERTY(QString containedInFolderPath   READ containedInFolderPath   WRITE setContainedInFolderPath)
 
 public:
    /**